+++ Kokoro/ModEntry.cs
@@ -58,6 +58,7 @@ public sealed class ModEntry : IModManifest, IApiProviderManifest, ISpriteManife
 		);
 
 		ActionCostsManager.Setup(Harmony);
+		ActionSourceManager.Setup(Harmony);
 		CardDestinationManager.Setup(Harmony);
 		CardRenderManager.Setup(Harmony);
 		ConditionalActionManager.Setup(Harmony);