+++ Kokoro/docs/release-notes.md
@@ -6,6 +6,7 @@
 
 * Kokoro now replaces all `shardcost` actions on the fly with its own action costs.
 * Extended the action costs system, allowing hooks to modify costs on the fly and add additional resource providers.
+* Added a new action source API.
 
 ## 2.1.1
 Released 8 January 2025.